+++ type/mvhd.go
@@ -96,8 +96,8 @@ func (b *MvhdBox) parse() (err error) {
 	duration := bmfcommon.DefaultEndianness.Uint32(data[16:20])
 
 	b.Standard32TimeSupport = bmfcommon.NewStandard32TimeSupport(
-		creationEpoch,
-		modificationEpoch,
+		uint64(creationEpoch),
+		uint64(modificationEpoch),
 		duration,
 		timeScale)
